Update solutions menu
