+++ pkg/model/solution/activesolution/creation.go
@@ -7,7 +7,9 @@ import (
 	"github.com/containerum/chkit/pkg/context"
 
 	"github.com/containerum/chkit/pkg/model/solution"
+	"github.com/containerum/chkit/pkg/porta"
 	"github.com/containerum/chkit/pkg/util/activekit"
+	"github.com/containerum/chkit/pkg/util/ferr"
 	"github.com/containerum/chkit/pkg/util/namegen"
 	"github.com/containerum/chkit/pkg/util/text"
 	"github.com/containerum/kube-client/pkg/model"
@@ -189,6 +191,19 @@ func Wizard(ctx *context.Context, config WizardConfig) solution.Solution {
 						}
 						border := strings.Repeat("_", text.Width(data))
 						fmt.Printf("%s\n%s\n%s\n", border, data, border)
+						return nil
+					},
+				},
+				{
+					Label: "Export solution to file",
+					Action: func() error {
+						var fname = activekit.Promt("Type filename: ")
+						fname = strings.TrimSpace(fname)
+						if fname != "" {
+							if err := (porta.Exporter{OutFile: fname}.Export(sol)); err != nil {
+								ferr.Printf("unable to export solution:\n%v\n", err)
+							}
+						}
 						return nil
 					},
 				},
